backend/requirements.txt
edgartools
fastapi
litellm
python-dotenv
uvicorn[standard]
